% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_areas.R
\name{calc_pixel_areas}
\alias{calc_pixel_areas}
\title{Calculates the pixel area for each line of a raster}
\usage{
calc_pixel_areas(x)
}
\arguments{
\item{x}{a \code{Raster*} object}
}
\value{
a vector with the area in square meters of the pixels in each line 
of \code{x} (vector has length equal to \code{nrow(x)})
}
\description{
Calculates the pixel area for each line of a raster
}
