% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_aoi.R
\name{check_aoi}
\alias{check_aoi}
\title{Check if aoi is an sf or sp type object
If sp, convert to sf and return}
\usage{
check_aoi(aoi)
}
\arguments{
\item{aoi}{the area of interest object}
}
\description{
Check if aoi is an sf or sp type object
If sp, convert to sf and return
}
