\name{gfc_animate}
\alias{gfc_animate}
\title{xxx}
\usage{
gfc_animate(aoi, output_path)
}
\arguments{
  \item{aoi}{an Area of Interest (AOI) as a
  \code{SpatialPolygons*} object. If the AOI is not in the
  WGS84 geographic coordinate system, it will be
  reprojected to WGS84.}
}
\description{
yyy
}
\examples{
#TODO: Add examples
}

