% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\encoding{UTF-8}
\name{test_poly}
\alias{test_poly}
\title{Polygon outlining TEAM site in Caxiuanã, Brazil}
\description{
Contains a \code{SpatialPolygonsDataFrame} with a simplified polygon of the
area within the Tropical Ecology Assessment and Monitoring (TEAM) network
site in Caxiuanã, Brazil.
}

