% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gfc.R
\name{extract_gfc}
\alias{extract_gfc}
\title{Extracts GFC data for a given AOI}
\usage{
extract_gfc(
  aoi,
  data_folder,
  to_UTM = FALSE,
  stack = "change",
  dataset = "GFC-2022-v1.10",
  ...
)
}
\arguments{
\item{aoi}{an Area of Interest (AOI) as a \code{sf} object.
If the AOI is not in WGS 1984 (EPSG:4326), it will be reprojected to WGS84.}

\item{data_folder}{folder where downloaded GFC product tiles are located 
(see \code{\link{download_tiles}} function.}

\item{to_UTM}{if TRUE, then reproject the output into the UTM zone of the 
AOI centroid. If FALSE, retain the original WGS84 projection of the GFC 
tiles.}

\item{stack}{the layers to extract from the GFC product. Defaults to 
"change". See Details.}

\item{dataset}{which version of the Hansen data to use}

\item{...}{additional arguments as for \code{\link{writeRaster}}, such as 
\code{filename}, or \code{overwrite}.}
}
\value{
\code{RasterStack} with GFC layers
}
\description{
This function extracts a dataset for a given AOI from a series of 
pre-downloaded GFC tiles. The \code{\link{download_tiles}} function should 
be used beforehand in order to download the necessary data to the specified
\code{data_folder}. Note that the output file format is fixed as GeoTIFF 
with LZW compression.
}
\details{
The \code{stack} option can be "change" (the default), "first", or "last".  
When set to "change", the forest change layers (treecover2000, loss, gain, 
lossyear, and datamask) will be extracted for the given \code{aoi}. The 
"first" and "last" options will mosaic the 2000 or last year composite top 
of atmosphere (TOA) reflectance images (respectively).
}
\seealso{
\code{\link{download_tiles}}, \code{\link{annual_stack}}, 
\code{\link{gfc_stats}}
}
