% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfcanalysis-package.R
\docType{data}
\encoding{UTF-8}
\name{gfc_tiles}
\alias{gfc_tiles}
\title{Grid of tiles used for the GFC product}
\description{
Contains a \code{SpatialPolygonsDataFrame} with a 10 x 10 degree grid in 
WGS84 coordinate system, covering the area from 180W-180E and 80N-60S, the 
tile system used by the GFC Product.
}
