% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfcanalysis-package.R
\docType{data}
\encoding{UTF-8}
\name{test_poly}
\alias{test_poly}
\title{Polygon outlining TEAM site in Caxiuanã, Brazil}
\description{
Contains a \code{SpatialPolygonsDataFrame} with a simplified polygon of the 
area within the Tropical Ecology Assessment and Monitoring (TEAM) network 
site in Caxiuanã, Brazil.
}
